API_Tours/requirements.txt
flask
flask-cors
mysqlclient
python-dotenv